enable suite averaging

+++ RoomDraw/db/migrate/20151018184508_create_draw_groups.rb
@@ -2,7 +2,7 @@ class CreateDrawGroups < ActiveRecord::Migration
   def change
     create_table :draw_groups do |t|
       t.integer :student_id
-      t.integer :draw_num
+      t.float :draw_num
       t.boolean :for_suite
       t.timestamps null: false
     end
